This is a program that counts instances of mantra submissions through Whatsapp. It is meant for Associação Buddha-Dharma associates in Brazil, so it only works in portuguese.
It reads the export text from Whatsapp line by line
and looks for strings that are in the format
"Fiz <number> mantras de <mantra name>."

It counts how many mantras, "Fiz", "mantras" and <mantra name> are read.
If there is a discrepancy in the counts it displays the line with the
date where it occured, so the person can edit the text to get an accurate
count.